Undo deprecation of Criterion type
This should be refactored in a different PR
