+++ src/Types/BooleanCriterion.php
@@ -4,13 +4,44 @@ namespace Distil\Types;
 
 use Distil\ActsAsCriteriaFactory;
 use Distil\Criterion;
-use Distil\Values\ConstructsFromBoolean;
+use Distil\Values\BooleanKeyword;
+use Distil\Values\ConstructsFromKeyword;
 
-/**
- * @deprecated 1.0.0 Use the according traits instead.
- */
 abstract class BooleanCriterion implements Criterion
 {
     use ActsAsCriteriaFactory;
-    use ConstructsFromBoolean;
+    use ConstructsFromKeyword;
+
+    public const KEYWORD_TRUE = 'true';
+    public const KEYWORD_FALSE = 'false';
+
+    private bool $value;
+
+    public function __construct(bool $value)
+    {
+        $this->value = $value;
+    }
+
+    /**
+     * @return static
+     */
+    public static function fromString(string $value): self
+    {
+        return self::fromKeyword(new BooleanKeyword($value));
+    }
+
+    public function value(): bool
+    {
+        return $this->value;
+    }
+
+    public function isTruthy(): bool
+    {
+        return $this->value;
+    }
+
+    public function isFalsy(): bool
+    {
+        return ! $this->value;
+    }
 }
