+++ src/Types/DateTimeCriterion.php
@@ -2,15 +2,50 @@
 
 namespace Distil\Types;
 
+use DateTimeInterface;
 use Distil\ActsAsCriteriaFactory;
 use Distil\Criterion;
-use Distil\Values\ConstructsFromDateTime;
+use Distil\Values\ConstructsFromKeyword;
+use Distil\Values\DateTimeKeyword;
 
-/**
- * @deprecated 1.0.0 Use the according traits instead.
- */
 abstract class DateTimeCriterion implements Criterion
 {
     use ActsAsCriteriaFactory;
-    use ConstructsFromDateTime;
+    use ConstructsFromKeyword;
+
+    private DateTimeInterface $value;
+    private string $format;
+
+    public function __construct(DateTimeInterface $value, string $format = DateTimeInterface::ATOM)
+    {
+        $this->value = $value;
+        $this->format = $format;
+    }
+
+    public function __toString(): string
+    {
+        if ($this->keyword) {
+            return (string) $this->keyword;
+        }
+
+        return $this->value->format($this->format);
+    }
+
+    /**
+     * @return static
+     */
+    public static function fromString(string $value): self
+    {
+        return self::fromKeyword(new DateTimeKeyword($value));
+    }
+
+    public function value(): DateTimeInterface
+    {
+        return $this->value;
+    }
+
+    public function format(): string
+    {
+        return $this->format;
+    }
 }
