+++ src/Types/StringCriterion.php
@@ -6,9 +6,6 @@ use Distil\ActsAsCriteriaFactory;
 use Distil\Criterion;
 use Distil\Values\ConstructsFromKeyword;
 
-/**
- * @deprecated 1.0.0 Use the according traits instead.
- */
 abstract class StringCriterion implements Criterion
 {
     use ActsAsCriteriaFactory;
