--- src/Values/ConstructsFromDateTime.php
@@ -1,46 +1,0 @@
-<?php
-
-namespace Distil\Values;
-
-use DateTimeInterface;
-
-trait ConstructsFromDateTime
-{
-    use ConstructsFromKeyword;
-
-    private DateTimeInterface $value;
-    private string $format;
-
-    public function __construct(DateTimeInterface $value, string $format = DateTimeInterface::ATOM)
-    {
-        $this->value = $value;
-        $this->format = $format;
-    }
-
-    public function __toString(): string
-    {
-        if ($this->keyword) {
-            return (string) $this->keyword;
-        }
-
-        return $this->value->format($this->format);
-    }
-
-    /**
-     * @return static
-     */
-    public static function fromString(string $value): self
-    {
-        return self::fromKeyword(new DateTimeKeyword($value));
-    }
-
-    public function value(): DateTimeInterface
-    {
-        return $this->value;
-    }
-
-    public function format(): string
-    {
-        return $this->format;
-    }
-}
